Hacked around a bunch in recovery code stuff
